Joggo/JoggoGames-Development-Primera-implementaci-n-de-Yo-nunca/requirements.txt
uvicorn[standard]==0.23.2
httpx==0.24.1
respx==0.20.1
requests==2.31.0
starlette-prometheus==0.10.0
mashumaro==3.5
pyhocon==0.3.60
aiosqlite==0.17.0
fastapi==0.95.2
pydantic==1.10.12
tortoise-orm==0.19.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4